pyproject.toml
[project]
name = "boardroom-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-firestore>=2.23.0",
    "google-generativeai>=0.8.6",
    "pypdf2>=3.0.1",
    "streamlit>=1.54.0",
]
